Sort out form components
